chore: rename to gsuite extractor
