+++ plugins/extractors/populate.go
@@ -10,8 +10,8 @@ import (
 	_ "github.com/odpf/meteor/plugins/extractors/elastic"
 	_ "github.com/odpf/meteor/plugins/extractors/gcs"
 	_ "github.com/odpf/meteor/plugins/extractors/github"
-	_ "github.com/odpf/meteor/plugins/extractors/googleworkspace"
 	_ "github.com/odpf/meteor/plugins/extractors/grafana"
+	_ "github.com/odpf/meteor/plugins/extractors/gsuite"
 	_ "github.com/odpf/meteor/plugins/extractors/kafka"
 	_ "github.com/odpf/meteor/plugins/extractors/mariadb"
 	_ "github.com/odpf/meteor/plugins/extractors/metabase"
